server/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic==2.11.3
pydantic-settings==2.8.1
python-dotenv==1.1.0
langchain==0.3.23
langchain_openai==0.3.12
langchain_community==0.3.21
langchain_core==0.3.51
langchain_pinecone==0.2.5
langchain-huggingface==0.1.2
langgraph==0.3.29
pinecone-client==6.0.0
openai==1.73.0
tavily-python==0.5.4
transformers==4.51.2